リフレクションでイベントインスタンスを作る部分を jvmMainに移動
